move constnats
